refactor: migrate and consolidate tests under `printer_service` module
Moved printer-related test cases from `test_app` and `test_print_auto_select` to `test_printer_service`. Updated tests to reflect modular changes in `printer_service`. Simplified mock setups and removed redundant `DeviceManager` references from `test_app`.
